+++ hooks/utils/utils.dart
@@ -1,0 +1,1 @@
+export 'gradle_updater.dart';
